+++ extensions/link-cleaner/src/main.ts
@@ -1,5 +1,6 @@
 import { Clipboard, showHUD } from "@raycast/api";
 import { rules } from "./rules";
+import { removeQueryParams } from "./utils";
 
 // detect all urls in text
 function findURLs(text: string): string[] {
@@ -11,28 +12,13 @@ function findURLs(text: string): string[] {
 // replace all urls in text
 function replaceURLs(text: string, newURLs: string[]): string {
   const urls = findURLs(text);
+  let modifiedText = text;
+
   for (const url of urls) {
-    text = text.replace(url, newURLs.shift() || url);
+    modifiedText = modifiedText.replace(url, newURLs.shift() || url);
   }
-  return text;
-}
 
-// remove some query params from url
-function removeQueryParams(url: string, allowParmas: string[]): string {
-  // find all query params
-  const urlParts = url.split("?");
-  if (urlParts.length < 2) {
-    return url;
-  }
-  const query = urlParts[1].split("&");
-
-  // if params is not empty, match params to remove
-  if (allowParmas.length > 0) {
-    const newQuery = query.filter((param) => allowParmas.includes(param.split("=")[0]));
-    return `${urlParts[0]}?${newQuery.join("&")}`;
-  }
-  // if params is empty, remove all query params
-  return urlParts[0];
+  return modifiedText;
 }
 
 export default async function main() {